Install Vue Router
