+++ package.json
@@ -5,9 +5,9 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
+    "test": "jest",
     "lint": "vue-cli-service lint",
-    "deploy": "bash deploy.sh",
-    "test": "jest"
+    "deploy": "bash deploy.sh"
   },
   "dependencies": {
     "bootstrap": "4.3.1",
@@ -16,6 +16,7 @@
     "shortid": "^2.2.14",
     "vue": "^2.6.10",
     "vue-click-outside": "1.0.7",
+    "vue-router": "^3.0.3",
     "vuex": "^3.0.1"
   },
   "devDependencies": {
