+++ src/components/pages/Home.vue
@@ -1,0 +1,3 @@
+<template>
+    <p>Suzuri is minimal markdown note for flash memo. <router-link to="/edit">Go to Editor</router-link></p> 
+</template>
