+++ src/router.ts
@@ -1,0 +1,31 @@
+import Vue from 'vue'
+import Router from 'vue-router'
+import Home from '@/components/pages/Home.vue';
+import Edit from '@/components/pages/Edit.vue';
+
+Vue.use(Router)
+
+export default new Router({
+  mode: 'history',
+  base: process.env.BASE_URL,
+  routes: [
+    {
+      path: '/',
+      name: 'home',
+      component: Home
+    },
+    {
+      path: '/edit',
+      name: 'edit',
+      component: Edit
+    },
+    {
+      path: '/about',
+      name: 'about',
+      // route level code-splitting
+      // this generates a separate chunk (about.[hash].js) for this route
+      // which is lazy-loaded when the route is visited.
+      component: () => import(/* webpackChunkName: "about" */ './views/About.vue')
+    }
+  ]
+})
